+++ lib/game.rb
@@ -1,12 +1,9 @@
 require_relative "hashable"
 require_relative "gameable"
-require_relative "winable"
-
 require_relative 'collection'
 
 class Game < Collection
 
-  include Winable
   extend Hashable
   extend Gameable
 
@@ -108,6 +105,13 @@ class Game < Collection
     @venue = game_stats[:venue]
     @venue_link = game_stats[:venue_link]
     @total_goals = @away_goals + @home_goals
+  end
+
+  def win?(team_id)
+    away_win = team_id == @away_team_id && @away_goals > @home_goals
+    home_win =  team_id == @home_team_id && @home_goals > @away_goals
+    return 1 if away_win || home_win
+    0
   end
 
 end
